gh #44 L3 test case development
Updated test for source devices
